+++ Hbuilder_Example/uni_modules/GC-UniPlugin/js_sdk/index.js
@@ -2,10 +2,6 @@ import {
 	gcActionTracking
 } from './Action/GCActionTracking.js';
 
-// #ifdef APP-HARMONY
-gcActionTracking.startTracking();
-// #endif
-
 export {
 	mobileAgent,
 	rum,